pyproject.toml
[tool.poetry]
name = "libre-api-pg"
version = "0.1.0"
description = ""
authors = ["Sam Purkis <sam.purkis@zimmerbiomet.com>"]
readme = "README.md"
packages = [{ include = "libre_api_pg" }]

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7,<4.0"
requests = "^2.31.0"
pandas = "^2.0.2"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"
polars = "^0.18.4"
matplotlib = "^3.7.1"
plotly = "^5.15.0"
streamlit = "^1.25.0"
arrow = "^1.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
